fix(kv): keep valkey out of the pub/sub module graph
`pubsub.ts` reached for `resolveCacheProvider()` in `clients.ts`, and
`clients.ts` statically imports every adapter — including `@keyv/valkey`,
which pulls in `iovalkey`. iovalkey uses a dynamic `require` of a file path,
so once `apps/service` imported `@chia/kv/pubsub` the bundled dev server
threw at request time:

    Dynamic require of ".../iovalkey/built/Redis.js" is not supported

`nitro build` did not surface it: rollup tree-shakes the unreachable adapter
away because nothing in the service calls `createKeyv()`. The dev pipeline
transforms modules individually and keeps the import, so only the running
server failed. Reproduced directly with esbuild — bundling a module that
imports `@chia/kv/pubsub` pulled in 69 iovalkey references before this
change and none after.

Move the provider resolution to its own module that depends on nothing but
`env`, so asking *which* backend is configured never drags in a client for
it. `clients.ts` re-exports it, so the seam is unchanged for callers.
