+++ packages/kv/src/clients.ts
@@ -1,48 +1,14 @@
 import type { Keyv } from "keyv";
 
-import { isUrl } from "@chia/utils/is";
-
 import { createPostgresKv } from "./adapters/postgres.ts";
 import { createRedisKv } from "./adapters/redis.ts";
 import { createValkeyKv } from "./adapters/valkey.ts";
 import { env } from "./env.ts";
+import { resolveCacheProvider } from "./provider.ts";
+
+export { resolveCacheProvider, type CacheProvider } from "./provider.ts";
 
 let kv: Keyv | null = null;
-
-/** Backends a resolved cache configuration can land on. `upstash` is REST-only. */
-export type CacheProvider = "redis" | "valkey" | "postgres" | "upstash";
-
-/**
- * Resolves the configured provider, following `auto` through to the URI's protocol.
- *
- * Split out from {@link createKeyv} so callers that need a *different* client for the same backend
- * — pub/sub needs its own connections, since a subscriber cannot run ordinary commands — can ask
- * what the backend is without a second copy of the protocol sniffing.
- */
-export const resolveCacheProvider = (): CacheProvider => {
-  if (env.CACHE_PROVIDER !== "auto") return env.CACHE_PROVIDER;
-
-  const protocol =
-    env.CACHE_URI &&
-    isUrl(env.CACHE_URI, {
-      allowedProtocols: ["redis", "valkey", "rediss", "valkeys", "postgres"],
-    })
-      ? new URL(env.CACHE_URI).protocol.replace(":", "")
-      : null;
-
-  switch (protocol) {
-    case "rediss":
-    case "redis":
-      return "redis";
-    case "valkeys":
-    case "valkey":
-      return "valkey";
-    case "postgres":
-      return "postgres";
-    default:
-      throw new Error(`Unsupported protocol: ${protocol}`);
-  }
-};
 
 export const createKeyv = () => {
   if (kv) {
