+++ packages/kv/src/provider.ts
@@ -1,0 +1,40 @@
+import { isUrl } from "@chia/utils/is";
+
+import { env } from "./env.ts";
+
+/**
+ * Which cache backend the environment resolves to, decided without touching a client.
+ *
+ * Deliberately its own module rather than living next to {@link createKeyv}: `clients.ts` statically
+ * imports every adapter, so anything importing it drags `@keyv/valkey` — and therefore `iovalkey`,
+ * which uses a dynamic `require` that bundlers refuse — into the graph. Callers that only need to
+ * *ask* what the backend is must not pay for that.
+ */
+
+/** Backends a resolved cache configuration can land on. `upstash` is REST-only. */
+export type CacheProvider = "redis" | "valkey" | "postgres" | "upstash";
+
+export const resolveCacheProvider = (): CacheProvider => {
+  if (env.CACHE_PROVIDER !== "auto") return env.CACHE_PROVIDER;
+
+  const protocol =
+    env.CACHE_URI &&
+    isUrl(env.CACHE_URI, {
+      allowedProtocols: ["redis", "valkey", "rediss", "valkeys", "postgres"],
+    })
+      ? new URL(env.CACHE_URI).protocol.replace(":", "")
+      : null;
+
+  switch (protocol) {
+    case "rediss":
+    case "redis":
+      return "redis";
+    case "valkeys":
+    case "valkey":
+      return "valkey";
+    case "postgres":
+      return "postgres";
+    default:
+      throw new Error(`Unsupported protocol: ${protocol}`);
+  }
+};
